requirements.txt
streamlit
cryptography
python-dotenv
sqlite3
